requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cssselect2==0.6.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2
django-allauth==0.51.0
django-amp-tools==0.2.2
django-ckeditor==6.4.1
django-classy-tags==3.0.1
django-cms==3.9.0
django-crispy-forms==1.14.0
django-embed-video==1.4.4
django-filer==2.2
django-formtools==2.3
django-import-export==2.8.0
django-ipware==4.0.2
django-js-asset==2.0.0
django-mptt==0.13.4
django-polymorphic==3.1.0
django-sekizai==3.0.1
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.1.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.8.1
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
lxml==4.8.0
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
reportlab==3.6.9
requests==2.28.1
requests-oauthlib==1.3.1
rq==1.10.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tablib==3.2.1
tinycss2==1.1.1
Unidecode==1.1.2
urllib3==1.26.10
webencodings==0.5.1
websockets==10.3
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
pandas~=1.5.3
